requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic[email]>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
